+++ src/main/java/core/mvc/view/RedirectView.java
@@ -1,30 +1,28 @@
-package core.mvc;
+package core.mvc.view;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Map;
 import java.util.Objects;
 
-public class ForwardView implements View {
+public class RedirectView implements View {
 
     private final String viewPath;
 
-    public ForwardView(String viewPath) {
+    public RedirectView(String viewPath) {
         this.viewPath = viewPath;
     }
 
     @Override
     public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        model.forEach(request::setAttribute);
-
-        request.getRequestDispatcher(viewPath).forward(request, response);
+        response.sendRedirect(viewPath);
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        ForwardView that = (ForwardView) o;
+        RedirectView that = (RedirectView) o;
         return Objects.equals(viewPath, that.viewPath);
     }
