+++ src/test/java/core/di/factory/example/MockView.java
@@ -1,6 +1,6 @@
 package core.di.factory.example;
 
-import core.mvc.View;
+import core.mvc.view.View;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -11,7 +11,5 @@ public class MockView implements View {
     @Override
     public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response)
             throws Exception {
-
     }
-
 }
