+++ src/test/java/next/controller/LoginControllerTest.java
@@ -1,7 +1,7 @@
 package next.controller;
 
-import core.mvc.ForwardView;
-import core.mvc.ModelAndView;
+import core.mvc.view.ForwardView;
+import core.mvc.view.ModelAndView;
 import core.mvc.asis.DispatcherServlet;
 import core.mvc.tobe.AnnotationHandlerMapping;
 import core.mvc.tobe.HandlerExecution;
